backend/requirements.txt
fastapi
uvicorn
pillow
numpy
joblib